fixed blog search
