+++ src/components/Features-carousel/index.js
@@ -52,10 +52,10 @@ const FeaturesCarousel = ({ features, heading }) => {
       <div className="features-carousel">
         <h2 className="main-heading">{heading ? heading : "Features"}</h2>
         <Slider
-          autoplay={true} 
-          autoplaySpeed={3500} 
-          arrows={false} 
-          dots={true} 
+          autoplay={true}
+          autoplaySpeed={3500}
+          arrows={false}
+          dots={true}
           infinite= {true}
           speed="500"
           slidesToShow={1}
