+++ src/sections/Blog/Blog-grid/index.js
@@ -15,21 +15,27 @@ const BlogGrid = ({
   setListView,
   setGridView,
   pageContext,
-  data,
+  searchData,
+  searchQuery,
+  setCurrentPage,
+  currentPage,
+  searchedPosts,
+  queryResults,
+  postsPerPage,
 }) => {
-  const [currentPage, setCurrentPage] = useState(1);
-  const [postsPerPage] = useState(10);
-  const indexOfLastPost = currentPage * postsPerPage;
-  const indexOfFirstPost = indexOfLastPost - postsPerPage;
-  const [searchQuery, setSearchQuery] = useState("");
-  const { queryResults, searchData } = useDataList(
-    data.allMdx.nodes,
-    setSearchQuery,
-    searchQuery,
-    ["frontmatter", "title"],
-    "id"
-  );
-  const searchedPosts = queryResults.slice(indexOfFirstPost, indexOfLastPost);
+  // const [currentPage, setCurrentPage] = useState(1);
+  // const [postsPerPage] = useState(10);
+  // const indexOfLastPost = currentPage * postsPerPage;
+  // const indexOfFirstPost = indexOfLastPost - postsPerPage;
+  // const [searchQuery, setSearchQuery] = useState("");
+  // const { queryResults, searchData } = useDataList(
+  //   data.allMdx.nodes,
+  //   setSearchQuery,
+  //   searchQuery,
+  //   ["frontmatter", "title"],
+  //   "id"
+  // );
+  // const searchedPosts = queryResults.slice(indexOfFirstPost, indexOfLastPost);
   // Change page
   const paginate = (pageNumber) => {
     setCurrentPage(pageNumber);
