+++ src/sections/Blog/Blog-list/index.js
@@ -15,25 +15,32 @@ const BlogList = ({
   setListView,
   setGridView,
   pageContext,
-  data,
+  // data,
+  currentPage,
+  searchQuery,
+  searchData,
+  setCurrentPage,
+  queryResults,
+  postsPerPage,
+  searchedPosts,
 }) => {
-  const [currentPage, setCurrentPage] = useState(1);
-  const [postsPerPage] = useState(10);
-  const indexOfLastPost = currentPage * postsPerPage;
-  const indexOfFirstPost = indexOfLastPost - postsPerPage;
-  let { totalCount, nodes } = data.allMdx;
-  const [searchQuery, setSearchQuery] = useState("");
-  const { queryResults, searchData } = useDataList(
-    nodes,
-    setSearchQuery,
-    searchQuery,
-    ["frontmatter", "title"],
-    "id"
-  );
+  // const [currentPage, setCurrentPage] = useState(1);
+  // const [postsPerPage] = useState(10);
+  // const indexOfLastPost = currentPage * postsPerPage;
+  // const indexOfFirstPost = indexOfLastPost - postsPerPage;
+  // let { totalCount, nodes } = data.allMdx;
+  // const [searchQuery, setSearchQuery] = useState("");
+  // const { queryResults, searchData } = useDataList(
+  //   nodes,
+  //   setSearchQuery,
+  //   searchQuery,
+  //   ["frontmatter", "title"],
+  //   "id"
+  // );
   const category = pageContext.category ? pageContext.category : null;
   const tag = pageContext.tag ? pageContext.tag : null;
-  const currentPosts = queryResults.slice(indexOfFirstPost, indexOfLastPost);
-  totalCount = queryResults.length;
+  // const searchedPosts = queryResults.slice(indexOfFirstPost, indexOfLastPost);
+  const totalCount = queryResults.length;
   const header = tag
     ? `${totalCount} post${totalCount === 1 ? "" : "s"} tagged with "${tag}"`
     : category
@@ -77,14 +84,14 @@ const BlogList = ({
               )}
               <div className="blog-list-wrapper">
                 <Row className="blog-lists">
-                  {currentPosts.length > 0 &&
-                    currentPosts.map(({ id, frontmatter, fields }) => (
+                  {searchedPosts.length > 0 &&
+                    searchedPosts?.map(({ id, frontmatter, fields }) => (
                       <Col xs={12} key={id}>
                         <Card  frontmatter={frontmatter} fields={fields} />
                       </Col>
                     ))}
                   <Col>
-                    {currentPosts.length > 0 && (
+                    {searchedPosts.length > 0 && (
                       <Pagination
                         postsPerPage={postsPerPage}
                         totalPosts={queryResults.length}
