+++ src/templates/blog-category-list.js
@@ -1,5 +1,6 @@
 import React from "react";
-
+import { useState, useEffect } from "react";
+import useDataList from "../utils/usedataList";
 
 
 import SEO from "../components/seo";
@@ -53,13 +54,37 @@ export const query = graphql`
 
 const BlogListPage = ({ pageContext, data }) => {
 
-
+  const [currentPage, setCurrentPage] = useState(1);
+  const [postsPerPage] = useState(10);
+  const indexOfLastPost = currentPage * postsPerPage;
+  const indexOfFirstPost = indexOfLastPost - postsPerPage;
+  let { totalCount, nodes } = data.allMdx;
+  const [searchQuery, setSearchQuery] = useState("");
+  const { queryResults, searchData } = useDataList(
+    nodes,
+    setSearchQuery,
+    searchQuery,
+    ["frontmatter", "title"],
+    "id"
+  );
+  const searchedPosts = queryResults.slice(indexOfFirstPost, indexOfLastPost);
   return (
 
     <>
 
 
-      <BlogList data={data} pageContext={pageContext} />
+      <BlogList
+        data={data}
+        pageContext={pageContext}
+        searchedPosts={searchedPosts}
+        setCurrentPage={setCurrentPage}
+        postsPerPage={postsPerPage}
+        searchData={searchData}
+        searchQuery={searchQuery}
+        setSearchQuery={setSearchQuery}
+        currentPage={currentPage}
+        queryResults={queryResults}
+      />
 
     </>
